+++ src/components/Culture/style.ts
@@ -1,0 +1,31 @@
+import styled from "styled-components";
+
+export const Culture = styled.div`
+    position: relative;
+
+    margin: 260px 0px 260px 0px;
+
+    width: 100vw;
+    height: 100vh;
+`;
+
+export const Container = styled.div`
+    position: absolute;
+
+    left: 10%;
+    width: 62.5%;
+    height: 100%;
+`;
+
+export const SubExplains = styled.div`
+    margin-top: 6rem;
+    position: absolute;
+
+    display: flex;
+    flex-direction: row;
+
+    justify-content: space-around;
+    align-items: center;
+`;
+
+
