--- src/components/Team-culture/style.ts
@@ -1,28 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-    position: relative;
-
-    display: flex;
-    flex-direction: column;
-
-    height: 493px;
-    width: 1200px;
-
-    padding-bottom: 85px;
-`;
-
-export const SubExplains = styled.div`
-    position: absolute;
-
-    display: flex;
-    flex-direction: row;
-
-    height: 100vh;
-    width: 100vw;
-
-    justify-content: space-around;
-    align-items: center;    
-`;
-
-
