+++ src/components/index.tsx
@@ -1,2 +1,3 @@
 export { default as Header } from './Header/index';
 export { default as Main } from './Main/index';
+export { default as Culture } from './Culture/index';
