+++ src/pages/Culture.tsx
@@ -1,0 +1,5 @@
+import CulturePage from '../components/Culture/index';
+
+export default function Culture() {
+  return <CulturePage />;
+}
