+++ src/pages/index.tsx
@@ -1,1 +1,2 @@
 export { default as Main } from './Main';
+export { default as Culture } from './Culture';
